Add CredentialManagementService and fix user ID validation
