+++ components/org.wso2.carbon.identity.api.user.credential/org.wso2.carbon.identity.api.user.credential.v2/src/main/java/org/wso2/carbon/identity/rest/api/user/credential/v2/utils/CredentialMgtEndpointUtils.java
@@ -24,20 +24,9 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.wso2.carbon.identity.api.user.common.error.APIError;
 import org.wso2.carbon.identity.api.user.common.error.ErrorDTO;
-import org.wso2.carbon.identity.api.user.credential.common.CredentialHandler;
-import org.wso2.carbon.identity.api.user.credential.common.CredentialManagementConstants.CredentialTypes;
 import org.wso2.carbon.identity.api.user.credential.common.CredentialManagementConstants.ErrorMessages;
-import org.wso2.carbon.identity.api.user.credential.common.dto.CredentialDTO;
-import org.wso2.carbon.identity.api.user.credential.common.dto.CredentialGroupDTO;
 import org.wso2.carbon.identity.api.user.credential.common.exception.CredentialMgtClientException;
 import org.wso2.carbon.identity.api.user.credential.common.exception.CredentialMgtException;
-import org.wso2.carbon.identity.rest.api.user.credential.v2.dto.CredentialEntryDTO;
-import org.wso2.carbon.identity.rest.api.user.credential.v2.dto.CredentialsByTypeDTO;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import javax.ws.rs.core.Response;
 
@@ -114,69 +103,5 @@ public class CredentialMgtEndpointUtils {
         error.setMessage(errorMessage);
         error.setDescription(errorDescription);
         return error;
-    }
-
-    /**
-     * Validates that a credential ID is not blank.
-     *
-     * @param credentialId Credential ID.
-     * @throws CredentialMgtClientException If the credential ID is blank.
-     */
-    public static void validateCredentialId(String credentialId) throws CredentialMgtClientException {
-
-        if (StringUtils.isBlank(credentialId)) {
-            throw new CredentialMgtClientException(ErrorMessages.ERROR_CODE_INVALID_CREDENTIAL_ID);
-        }
-    }
-
-    /**
-     * Builds a CredentialsByTypeDTO by iterating over all registered credential handlers and populating
-     * the response fields for each configured credential type.
-     *
-     * @param handlerMap Map of credential types to their handlers.
-     * @param userId     User ID for whom credentials are retrieved.
-     * @return Populated CredentialsByTypeDTO.
-     */
-    public static CredentialsByTypeDTO buildCredentialsByTypeDTO(
-            Map<CredentialTypes, CredentialHandler> handlerMap, String userId) {
-
-        CredentialsByTypeDTO response = new CredentialsByTypeDTO();
-        for (Map.Entry<CredentialTypes, CredentialHandler> entry : handlerMap.entrySet()) {
-            try {
-                CredentialGroupDTO group = entry.getValue().getCredentials(userId);
-                if (!group.isConfigured()) {
-                    continue;
-                }
-                switch (entry.getKey()) {
-                    case PASSKEY:
-                        response.passkey(toEntryDTOs(group.getCredentials()));
-                        break;
-                    case PUSH_AUTH:
-                        response.pushAuth(toEntryDTOs(group.getCredentials()));
-                        break;
-                    case BACKUP_CODE:
-                        response.backupCode(true);
-                        break;
-                    default:
-                        LOG.warn("No response mapping for credential type: " + entry.getKey() + ". Skipping.");
-                        break;
-                }
-            } catch (CredentialMgtException e) {
-                throw handleCredentialMgtException(e);
-            }
-        }
-        return response;
-    }
-
-    private static List<CredentialEntryDTO> toEntryDTOs(List<CredentialDTO> credentials) {
-
-        if (credentials == null) {
-            return Collections.emptyList();
-        }
-        return credentials.stream()
-                .map(dto -> new CredentialEntryDTO()
-                        .credentialId(dto.getCredentialId())
-                        .displayName(dto.getDisplayName()))
-                .collect(Collectors.toList());
     }
 }
